+++ Assets/Scripts/Drinks/DrinkSpawner.cs
@@ -12,7 +12,6 @@ public class DrinkSpawner : MonoBehaviour
     // Update is called once per frame
     void OnCollisionEnter2D(Collision2D collision)
     {
-
         if (collision.gameObject.CompareTag("Player"))
         {
             Drink.SetActive(true);
